+++ src/components/Service.jsx
@@ -10,8 +10,8 @@ import LowerPageTop from './atoms/LowerPageTop';
 import TopicPath from './atoms/TopicPath';
 import Blog from './atoms/Blog';
 import Contact from './atoms/Contact';
-import ScrollToTop from './atoms/scrollToTop';
-import Index from './atoms/Index';
+import ScrollToTop from './atoms/ScrollToTop';
+import Table from './atoms/Table';
 
 const Service = () => {
 
@@ -289,7 +289,7 @@ const Service = () => {
         <ScrollToTop />
       </MediaQuery>
       <MediaQuery query='(max-width: 767px)'>
-        <Index 
+        <Table
           contents={ contents }
         />
       </MediaQuery>
