+++ src/components/atoms/Button.jsx
@@ -2,7 +2,6 @@
 
 import React from 'react';
 import { Link } from 'react-router-dom';
-import 'aos/dist/aos.css';
 
 const Button = (props) => {
